+++ www/models.py
@@ -2,6 +2,7 @@ import datetime
 
 from sqlalchemy.schema import Column
 from sqlalchemy.types import Integer
+from sqlalchemy.types import Float
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -9,10 +10,10 @@ Base = declarative_base()
 class Samples(Base):
     __tablename__ = 'samples'
     id=Column(Integer, primary_key=True)
-    temperature=Column('temperature', Integer)
-    humidity=Column('humidity', Integer)
-    pressure=Column('pressure', Integer)
-    windspeed=Column('windspeed', Integer)
+    temperature=Column('temperature', Float)
+    humidity=Column('humidity', Float)
+    pressure=Column('pressure', Float)
+    windspeed=Column('windspeed', Float)
     
     def serialize(self):
     		return {
